moved tests for simplicity
